Add a tool definition for the Numpy C API.
While the C API requires a little more in the way of gymnastics
than is usual for pybindings, it makes passing multidimensional
arrays in and out of Python sooooo much less painful. We don't
actually want to make Numpy a required library, so projects using
this in CMakeLists.txt should wrap everything in if(NUMPY_FOUND).


git-svn-id: http://code.icecube.wisc.edu/svn/highlander/trunk@169462 16731396-06f5-0310-8873-f7f720988828