Catch log_fatal() inside of property getters when checking whether the property exists (via leif).
Backstory: hasattr() works by calling getattr(), and returning False if
any error is raised. If a getter (say, for the mass of a particle that has
no defined mass) raises an error, the dynamism-disabling __getattr__()
shim will raise its own error rather than the underlying one, which is
quite confusing. Instead, use a modified hasattr() that checks the error
type explicitly, and returns True if RuntimeError is raised.



git-svn-id: http://code.icecube.wisc.edu/svn/highlander/trunk@169866 16731396-06f5-0310-8873-f7f720988828