relax the forcing of cmake variables
The gratuitous use of set(... FORCE) causes changes deep in the dependency
tree. This triggers a "touch" of files that the Makefile generator uses,
which in turn causes make to think that projects are out of date, forcing
a rebuild when it's not necessary.

I was able to reproduce this consistantly in two rare situations, and others
have seen this behavior as well. This patch at least fixes the two situations
I observed.


git-svn-id: http://code.icecube.wisc.edu/svn/highlander/trunk@170070 16731396-06f5-0310-8873-f7f720988828