fix the default warning suppression flag tests
GCC doesn't flag unknown warning suppression flags at low warning levels.
you can litterally use "-Wno-poop" and GCC will ignore it. this obviously
doesn't work for testing single flags. So, we change the test logic to look
for the warning ("-Wpoop") if that exists, then we enable the suppression,
because we assume that suppression doesn't exist without it's corresponding
warning.


git-svn-id: http://code.icecube.wisc.edu/svn/highlander/trunk@170188 16731396-06f5-0310-8873-f7f720988828