add a script to produce per-project coverage reports
Running this script with a project name will produce a coverge report for
that particular projec. Some ancilary data from other projects may be pulled
in, but that can be ignored by the reader, as it will be incomplete.

An execution example would be: 

  $ cd $I3_BUILD
  $ ./env-shell.sh bin/coverage.sh dataio

This will produce a report for the dataio project in ${I3_BUILD}/docs/coverage


git-svn-id: http://code.icecube.wisc.edu/svn/highlander/trunk@170265 16731396-06f5-0310-8873-f7f720988828