Recursively cache parents of parents when merging frames
This fixes a severe performance regression introduced in r149410, which
attempted to fix #1719. As a side-effect, though, it only cached direct
parents. This meant that the Q parent of a P frame would be cached, but GCD
frames would be read from the input file for every P frame. For a gzipped
Upgrade mDOM GCD, this caused steamshovel to take 30 seconds to move between
frames.


git-svn-id: http://code.icecube.wisc.edu/svn/highlander/trunk@170521 16731396-06f5-0310-8873-f7f720988828