Fix an initialization bug in bayesian blocks rebinning
Previously the rebinning ran from the second to the second-to-last bin
in an ill-considered attempt to preserve the bins before the first and
after the last pulse. Since `last` was initialized with 0, however,
the change-point-peeling loop terminated immediately upon encountering
`last[raw_bins-1] == 0`. This effectively made BinSigma equivalent
to PhotonsPerBin=-1, resulting in correpsondingly terrible angular
resolutions. Hat tip to Zelong Zhang at SBU for finding this 1-line
change.

We fix this by abandoning the idea of preserving the first and last
bins. This is not all that useful in general, and actively harmful if
those bins butt up against an exclusion window.


git-svn-id: http://code.icecube.wisc.edu/svn/meta-projects/combo/trunk@171365 16731396-06f5-0310-8873-f7f720988828