Add kernel support and benchmarks for counter-based RNGs from Random123
Counter-based RNGs generate random numbers by applying a
pseudo-cryptographic hash to a simple counter. As such, the state is
small and can be trivially initialized, partitioned, and advanced. This
means that it can be attached to the work item rather than the thread,
making the kernel stateless and the result strongly reproducible. The
implementations in Random123 pass all the tests in BigCrush, and some are
only ~3x slower than the dead-simple (but stateful) MWC we currently use.



git-svn-id: http://code.icecube.wisc.edu/svn/meta-projects/combo/trunk@171636 16731396-06f5-0310-8873-f7f720988828