Fix photospline_hitmaker_test by using a consistent set of tracks
The number of PE is an important quantity to track, but it depends
strongly on the distance of closest approach to a DOM. Since the tracks
in this test were drawn randomly, the set of tracks tested depends on how
many random numbers are consumed between frames by e.g. PROPOSAL. To
"fix" this, we hard-code the set of tracks emitted by the test in
photonics-service V01-02-05, where it was known to pass.



git-svn-id: http://code.icecube.wisc.edu/svn/meta-projects/combo/trunk@172164 16731396-06f5-0310-8873-f7f720988828