Fix #2136.
When adding tray segments, I3Tray automatically propagates If into the
modules added by the segment. However, it used a single global variable
that could be overridden if that segment added another segment with a
different condition (or that had a keyword argument named "If"), resulting
in generalized space madness. Fix this by maintain a depth-indexed record
of conditions.


git-svn-id: http://code.icecube.wisc.edu/svn/meta-projects/combo/trunk@172193 16731396-06f5-0310-8873-f7f720988828