Disable native math on CPUs; it is super dangerous
native_rsqrt(float) resolves to rsqrtss on x86, which has only 11 bit
precision. This is enough to destroy the normalization of a photon
direction such that the spherical intersection calculation (which assumes
proper normalization) finds an intersection with a DOM 20 m away, but
the actual photon path misses by a full diameter. The things you learn.



git-svn-id: http://code.icecube.wisc.edu/svn/meta-projects/combo/trunk@172734 16731396-06f5-0310-8873-f7f720988828