Reduce busy-waiting by polling for separately for output
If you poll with POLLOUT|POLLIN while only one is true, but require both
to process, you end up with a busy loop of zmq::poll(). This caused
CPU usage to shoot up to 200% while the server processed a block of
steps. This is the opposite of what was intended.

Also, remove superfluous error check on the output of zmq::poll(). The
cxx bindings wrapper throws an exception if zmq_poll() returns an error
code, so the check will never be reached.



git-svn-id: http://code.icecube.wisc.edu/svn/meta-projects/combo/trunk@175178 16731396-06f5-0310-8873-f7f720988828