Properly emulate the output of I3MuonSlicer in I3PropagatorServicePROPOSAL
While emitting ContinuousEnergyLoss segments seems attractive, the
spectrum of secondary tracks they represent (and thus the Cherenkov
yield) depends on the underlying track energy, the stochastic cut, and
the the total energy lost. Since the existing yield parameterization
from https://arxiv.org/abs/1206.5530 depends only on the first two,
it is simpler to stick to that form.

This also fixes a silent regression in which
icecube.simprod.segments.PropagateMuons.make_standard_propagators ignored
the EmitTrackSegments kwarg, causing I3CLSimMakePhotonsWithServer to
attempt to propagate photons by the track outside the detector volume. For
100 TeV slightly upgoing tracks, these escaped the atmosphere and could
be 1e7 m long.



git-svn-id: http://code.icecube.wisc.edu/svn/meta-projects/combo/trunk@176190 16731396-06f5-0310-8873-f7f720988828