Account for cable shadow in propagation
This implementation checks whether a photon path that would intersect a
DOM also intersects the cable. If so, no hit is recorded. The cable
position and radius is taken from an I3Map<ModuleKey,
I3ExtraGeometryItemCylinder> created by GetIceCubeCableShadow() and put
in the Geometry frame as "CableShadow". All cylinders must have the same
radius and rest on the surface of the DOM; the orientation of the cable
is discretized to 360/256 degrees, or 10% of a cable diameter. The cable
shadow is enabled by default.

See #2388.

git-svn-id: http://code.icecube.wisc.edu/svn/meta-projects/combo/trunk@178169 16731396-06f5-0310-8873-f7f720988828