Enable GNU extensions in xppc
This allows dot-notation access to fields of cl_floatN.

-std=c++11 enables __STRICT_ANSI__, -std=gnu++11 does not. Strict ANSI
mode disables, among other things, the union aliases xyzw for the fields
of cl_float4, as the C standard only requires that arrays be contiguous,
but allows implementations to choose any struct padding scheme they
please. This means that a compliant compiler can choose a different
memory layout for cl_float4 in C than float4 in OpenCL if the xyzw
aliases were used.

git-svn-id: http://code.icecube.wisc.edu/svn/meta-projects/combo/trunk@180137 16731396-06f5-0310-8873-f7f720988828